Refactor: Isolating components for tpye errors #84
